main payment product_board context 구조 수정 - 12/23 강민
