+++ src/main/moivo_react/src/App.jsx
@@ -33,6 +33,8 @@ import KakaoCallback from './components/kakao/KakaoCallback';
 import Admins_FAQ from './containers/admin/admins_FAQ';
 import Admins_ProductList from './containers/admin/admin_productList';
 import ReviewWrite from './containers/review/review_write';
+import DashBoardProvider from './contexts/DashBoardContext';
+import PaymentProvider from './contexts/payment/PaymentContext';
 
 
 
@@ -77,11 +79,15 @@ const App = () => {
     <Router>
       <AuthProvider>
         <MainProvider>
-          <Routes>
-            {routeConfig.map(({ path, element }, index) => (
-              <Route key={index} path={path} element={element} />
-            ))}
-          </Routes>
+          <DashBoardProvider>
+            <PaymentProvider>
+              <Routes>
+                {routeConfig.map(({ path, element }, index) => (
+                  <Route key={index} path={path} element={element} />
+                ))}
+              </Routes>
+            </PaymentProvider>
+          </DashBoardProvider>
         </MainProvider>
       </AuthProvider>
     </Router>
